wedding_planner/get_venues/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
pandas==2.*
db-dtypes==1.*